chore(ADE-495): Refactor code and cleanup atlassian-plugin.xml
