+++ src/main/java/velocity_implicit.vm
@@ -1,3 +1,3 @@
 #* @implicitly included *#
 
-#* @vtlvariable name="message" type="au.com.agiledigital.idea_search.KeyProperty" *#
+#* @vtlvariable name="message" type="au.com.agiledigital.idea_search.blueprints.KeyProperty" *#
