+++ src/main/resources/ideaSearch.properties
@@ -1,16 +1,18 @@
-ideaSearch.blueprint.name=Idea Search
-ideaSearch.blueprint.title=Idea Search
-ideaSearch.create-link.title=Fedex idea
-ideaSearch.create-link.description=Add a fedex idea to the list
-ideaSearch.blueprint.wizard.page1.title=Share your idea
-ideaSearch.blueprint.wizard.page1.desc.header=About this form
-ideaSearch.blueprint.wizard.page1.desc.content=Use this to create a page with your name and information about your idea
-ideaSearch.blueprint.form.label.idea.title=Add a name
-ideaSearch.blueprint.form.label.idea.description=Add a description
-ideaSearch.blueprint.form.label.idea.owner=Who had the idea?
-ideaSearch.blueprint.form.label.idea.status=Add a status
-ideaSearch.blueprint.form.label.idea.team=Add other people
+ideaSearch.blueprint.page.name=Idea Search
+ideaSearch.blueprint.page.title=Idea Search
+ideaSearch.blueprint.page.create-link.title=Fedex idea
+ideaSearch.blueprint.page.create-link.description=Add a fedex idea to the list
+ideaSearch.blueprint.page.wizard.page1.title=Share your idea
+ideaSearch.blueprint.page.wizard.page1.desc.header=About this form
+ideaSearch.blueprint.page.wizard.page1.desc.content=Use this to create a page with your name and information about your idea
+ideaSearch.blueprint.page.form.label.idea.title=Add a name
+ideaSearch.blueprint.page.form.label.idea.description=Add a description
+ideaSearch.blueprint.page.form.label.idea.owner=Who had the idea?
+ideaSearch.blueprint.page.form.label.idea.status=Add a status
+ideaSearch.blueprint.page.form.label.idea.team=Add other people
 ideaSearch.macro.client-side-scripts.desc=Client Side Script Macro for importing scripts for Blueprint
 ideaSearch.macro.structuredField.desc=Field containing structured data for the idea
+
+# Macro wizard
 au.com.agiledigital.idea_search.ideaSearch-macro-structuredField.param.category.label=Category:
 au.com.agiledigital.idea_search.ideaSearch-macro-structuredField.param.category.desc=Pick a category for the structured data
