+++ src/main/resources/soy/ideaSearch.soy
@@ -1,4 +1,4 @@
-{namespace IdeaSearch.Blueprints.FedexIdea}
+{namespace ideaSearch.blueprint.page.fedexIdea}
 
 /**
 *  A form that accepts the basic information for an idea
@@ -7,7 +7,7 @@
   <form action="#" method="post" class="aui">
     <fieldset>
       <div class="field-group">
-        <label for="videatitle">{getText('ideaSearch.blueprint.form.label.idea.title')}</label>
+        <label for="videatitle">{getText('ideaSearch.blueprint.page.form.label.idea.title')}</label>
         <input id="videatitle" class="text" type="text"
                name="vIdeaTitle">*
       </div>
@@ -15,21 +15,21 @@
     <fieldset>
       <div class="field-group">
         <label
-            for="videadescription">{getText('ideaSearch.blueprint.form.label.idea.description')}</label>
+            for="videadescription">{getText('ideaSearch.blueprint.page.form.label.idea.description')}</label>
         <input id="videadescription" class="text" type="text"
                name="vIdeaDescription"/>
       </div>
     </fieldset>
     <fieldset>
       <div class="field-group">
-        <label for="videaowner">{getText('ideaSearch.blueprint.form.label.idea.owner')}</label>
+        <label for="videaowner">{getText('ideaSearch.blueprint.page.form.label.idea.owner')}</label>
         <input id="videaowner" class="text" type="text"
                name="vIdeaOwner"/>
       </div>
     </fieldset>
     <fieldset>
       <div class="field-group">
-        <label for="videastatus">{getText('ideaSearch.blueprint.form.label.idea.status')}</label>
+        <label for="videastatus">{getText('ideaSearch.blueprint.page.form.label.idea.status')}</label>
         <select id="videastatus" name="vIdeaStatus">
           <option value="new">New</option>
           <option value="inProgress">In progress</option>
@@ -39,7 +39,7 @@
     </fieldset>
     <fieldset>
       <div class="field-group">
-        <label for="videateam">{getText('ideaSearch.blueprint.form.label.idea.team')}</label>
+        <label for="videateam">{getText('ideaSearch.blueprint.page.form.label.idea.team')}</label>
         <input id="videateam" class="text" type="text"
                name="vIdeaTeam"/>
       </div>
