+++ src/main/resources/templates/pageBlueprint.xml
